Allow buckets to be deleted.

+++ lib/arriba/volume/s3_directory.rb
@@ -168,11 +168,15 @@ module Arriba
 
     def rm(path)
       src = S3Path.new(path)
+      if src.key == ""
+        # delete the bucket
+        target.get_bucket(src.bucket).destroy
+      else
         target.get_bucket(src.bucket, src.key).files.each { |object|
           # This list of objects will include the item we're trying to delete and all children
           object.destroy
         }
-
+      end
     end
     
     def mkdir(path, newdir)
